Fix hydration test setup and refactor hydration test data
