+++ src/data/user-test-data.js
@@ -24,60 +24,7 @@ const userTestData = [{
     24,
     19
   ]
-}]
+}];
 
-const hydrationTestData = [
-  {
-    userID: 2,
-    date: "2019/06/15",
-    numOunces: 75
-  },
-  {
-    userID: 2,
-    date: "2019/06/16",
-    numOunces: 91
-  },
-  {
-    userID: 1,
-    date: "2019/06/15",
-    numOunces: 37
-  },
-  {
-    userID: 1,
-    date: "2019/06/16",
-    numOunces: 69
-  },
-  {
-    userID: 1,
-    date: "2019/06/17",
-    numOunces: 96
-  },
-  {
-    userID: 1,
-    date: "2019/06/18",
-    numOunces: 61
-  },
-  {
-    userID: 1,
-    date: "2019/06/19",
-    numOunces: 91
-  },
-  {
-    userID: 1,
-    date: "2019/06/20",
-    numOunces: 50
-  },
-  {
-    userID: 1,
-    date: "2019/06/21",
-    numOunces: 50
-  },
-  {
-    userID: 1,
-    date: "2019/06/22",
-    numOunces: 43
-  }
-];
 
-export default hydrationTestData;
 export default userTestData;
